requirement.txt
streamlit
langchain==0.1.20
langchain-core==0.1.52
langchain-google-genai==1.0.3
google-generativeai
python-dotenv
requests